requirements.txt
# Core ML and Data Processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
tensorflow==2.15.0

# API and Web Framework
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pydantic==2.5.2
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Redis for caching
redis==5.0.1
celery==5.3.4

# Data Sources and APIs
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0

# F1 Data Sources
fastf1==3.2.0
ergast-py==1.0.1

# Weather APIs
pyowm==3.3.0

# Data Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging and Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Security
cryptography==41.0.8
bcrypt==4.1.2
